Struct ioapic::IoApic

pub struct IoApic {
    pub id: u8,
    /* private fields */
}

A representation of an IoApic (x86-specific interrupt chip for I/O devices).

Fields
id: u8

The ID of this IoApic.

Implementations

impl IoApic

pub fn new(
page_table: &mut PageTable,
id: u8,
phys_addr: PhysicalAddress,
gsi_base: u32
) -> Result<(), &'static str>

Creates a new IoApic struct from the given id, PhysicalAddress, and gsi_base, and then adds it to the system-wide list of all IOAPICs.

pub fn handles_irq(&self, irq_num: u32) -> bool

Returns whether this IoApic handles the given irq_num, i.e., whether it’s within the range of IRQs handled by this IoApic.

pub fn id(&mut self) -> u32

gets this IoApic’s id.

pub fn version(&mut self) -> u32

gets this IoApic’s version.

pub fn arbitration_id(&mut self) -> u32

gets this IoApic’s arbitration id.

pub fn mask_irq(&mut self, irq: u8)

Masks (disables) the given IRQ line. NOTE: this function is UNTESTED!

pub fn set_irq(&mut self, ioapic_irq: u8, lapic_id: u8, irq_vector: u8)

Set IRQ to an interrupt vector.

Arguments

ioapic_irq: the IRQ number that this interrupt will trigger on this IoApic.

lapic_id: the id of the LocalApic that should handle this interrupt.

irq_vector: the system-wide IRQ vector number, which after remapping is from 0x20 to 0x2F (see interrupts::IRQ_BASE_OFFSET). For example, 0x20 is the PIT timer, 0x21 is the PS2 keyboard, etc.